fix: move useObserver to hooks
