+++ src/modules/dashboard/components/ExpenditurePage/Stages/constants.ts
@@ -30,21 +30,3 @@ export interface Motion {
   type: MotionType;
   status: MotionStatus;
 }
-
-export const useObserver = (
-  setFieldErrorsAmount: React.Dispatch<React.SetStateAction<number>>,
-) => {
-  const observer = new MutationObserver((mutations) => {
-    mutations.forEach((mutation) => {
-      if (mutation.type === 'attributes') {
-        const nodeListLength = document
-          .getElementById('expenditurePage')
-          ?.querySelectorAll('[aria-invalid="true"]').length;
-
-        setFieldErrorsAmount(nodeListLength || 0);
-      }
-    });
-  });
-
-  return { observer };
-};
